old-tests/submission/W3C/domhtml/HTMLDlistElement01.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - DL</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/obsolete.html#dom-dl-compact" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<dl compact="COMPACT">
<dd>Accountant</dd>
<dd>56,000</dd>
<dd>Female</dd>
<dd>1230 North Ave. Dallas, Texas 98551</dd>
</dl>


<script>
var domtest = async_test("The compact IDL attribute of the dl element must reflect the content attribute of the same name.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vcompact;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("dl");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vcompact = testNode.compact;

      assert_true(vcompact);

    domtest.done();
}));

</script>
<p>The compact IDL attribute of the dl element must reflect the content attribute of the same name.</p>
<div id="log"></div>
</body></html>
